fancy makefiles, fixes for the various tools and fixed a missing #include
Makefiles now look like this:

    $ make
     CC      build/ruby/src/pokemon_1.o
    $ make V=1
    tools/preproc/preproc -I tools/agbcc/include -I include -DRUBY
    -DREVISION=0 -DENGLISH -DDEBUG=0 -c charmap.txt src/pokemon_1.c |
    tools/agbcc/bin/agbcc -mthumb-interwork -Wimplicit -Wparentheses
    -Wunused -Werror -O2 -fhex-asm |
    /Users/easyaspi/pokeruby2/tools/binutils/bin/arm-none-eabi-as
    -mcpu=arm7tdmi -I include --defsym RUBY=1 --defsym REVISION=0
    --defsym ENGLISH=1 --defsym DEBUG=0 -W -o build/ruby/src/pokemon_1.o

Makefiles will also avoid using bash when KEEP_TEMPS is enabled. That is
because we don't have any significant piping.

- scaninc now memory maps the files (using the MIO library) and uses
  better algorithms.
  - It uses unordered_set instead of set. Order really doesn't matter.
  - It also tries to use faster C++17 features (string_view,
    boyer_moore_horspool_searcher) if possible, falling back to C++11 if it
    can't.
  - C++-isms instead of C-isms. I think scaninc should stay C++ for now, unless
    we decide to merge it into preproc.
  - I have cracked down on code duplication, using a base template class for
    AsmFile and CFile.
- gbagfx now enforces strict const-correctness and changes a few lines to avoid
  unwanted int promotion during an expensive modulo operation. (rip ARM)
- preproc has had some minor tweaks.
  - ASM files no longer derp on -o.
  - CharmapReader_Delete no longer tries to dereference a null pointer.
  - khash uses floats instead of doubles because we don't need that precisison
  - Line numbers are inaccurate! I am still working that out, and would
    like some help.
    The thing this causes issues with is the error messages from agbcc.
  - Fixed a few formatting :morphon:s

+++ tools/preproc/c_file.c
@@ -421,7 +421,7 @@ static void TryConvertIncbin(CFile *m)
         if (isSigned)
             fprintf(g_file, "%d,", data);
         else
-            fprintf(g_file, "%uu,", data);
+            fprintf(g_file, "%uu,", (unsigned)data);
     }
 
     putc_unlocked('}', g_file);
